+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 This project uses [Gitmoji](https://gitmoji.carloscuesta.me/)
 
+## 4.2.3
+
+- :arrow_up: dependency upgrade.
+- :building_construction: improve types for `arrayFilterIn`.
+- :building_construction: improve types for `arrayFilterOut`.
+- :building_construction: improve types for `arrayFilterTuple`.
+
 ## 4.2.2
 
 - :arrow_up: dependency upgrade.